backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.0
chromadb==0.5.0
numpy<2
httpx==0.27.2
python-docx==1.1.2
pypdf==4.3.1
ragas==0.4.2
langchain==1.2.15
langchain-core==1.3.2
langchain-openai==1.2.1
langchain-chroma==1.1.0
openai>=1.55.0
instructor>=1.7.0
